fix: fix issue with tooltip flickering
The mouse over event bubbles up to parent elements, so even though the mouse out event fired to
close a previous tooltip, the mouse over event would bubble up and reactivate the tooltip. In the
best case this made duplicate tooltips, in the worst case it created visual glitches.

fix #20